Seperate Board class from clear_FIFO
Move Board class into VHF/board_init, some logging is also introduced.
